add a rule for PartialAggregatePushDown
