+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/v2/V2ScanRelationPushDown.scala
@@ -18,11 +18,9 @@
 package org.apache.spark.sql.execution.datasources.v2
 
 import org.apache.spark.sql.catalyst.expressions._
-import org.apache.spark.sql.catalyst.expressions.aggregate.AggregateExpression
 import org.apache.spark.sql.catalyst.planning.ScanOperation
-import org.apache.spark.sql.catalyst.plans.logical.{Aggregate, Filter, LogicalPlan, Project}
+import org.apache.spark.sql.catalyst.plans.logical.{Filter, LogicalPlan, Project}
 import org.apache.spark.sql.catalyst.rules.Rule
-import org.apache.spark.sql.catalyst.util.toPrettySQL
 import org.apache.spark.sql.connector.read.{Scan, V1Scan}
 import org.apache.spark.sql.execution.datasources.DataSourceStrategy
 import org.apache.spark.sql.sources
@@ -33,101 +31,6 @@ object V2ScanRelationPushDown extends Rule[LogicalPlan] with AliasHelper {
   import DataSourceV2Implicits._
 
   override def apply(plan: LogicalPlan): LogicalPlan = plan transformDown {
-    // Pattern matching for partial aggregate push down
-    // For parquet, footer only has statistics information for max/min/count.
-    // It doesn't handle max/min/count associated with filter or group by.
-    // ORC is similar. If JDBC partial aggregate push down is added later,
-    // these condition checks need to be changed.
-    case aggNode@Aggregate(groupingExpressions, resultExpressions, child)
-      if (groupingExpressions.isEmpty) =>
-      child match {
-        case ScanOperation(project, filters, relation: DataSourceV2Relation)
-          if (filters.isEmpty) =>
-          val scanBuilder = relation.table.asReadable.newScanBuilder(relation.options)
-          var aggregates = resultExpressions.flatMap { expr =>
-            expr.collect {
-              case agg: AggregateExpression =>
-                replaceAlias(agg, getAliasMap(project)).asInstanceOf[AggregateExpression]
-            }
-          }
-          aggregates = DataSourceStrategy.normalizeExprs(aggregates, relation.output)
-              .asInstanceOf[Seq[AggregateExpression]]
-
-          val aggregation = PushDownUtils
-              .pushAggregates(scanBuilder, aggregates, groupingExpressions)
-          if (aggregation.aggregateExpressions.isEmpty) {
-            aggNode // return original plan node
-          } else {
-            // use the aggregate columns as the output columns
-            // e.g. TABLE t (c1 INT, c2 INT, c3 INT)
-            // SELECT min(c1), max(c1) FROM t;
-            // Use min(c1), max(c1) as output for DataSourceV2ScanRelation
-            // We want to have the following logical plan:
-            // == Optimized Logical Plan ==
-            // Aggregate [min(min(c1)#21) AS min(c1)#17, max(max(c1)#22) AS max(c1)#18]
-            // +- RelationV2[min(c1)#21, max(c1)#22] parquet file ...
-            val output = aggregates.map {
-              case agg: AggregateExpression =>
-                AttributeReference(toPrettySQL(agg), agg.dataType)()
-            }
-
-            // No need to do column pruning because only the aggregate columns are used as
-            // DataSourceV2ScanRelation output columns. All the other columns are not
-            // included in the output. Since PushDownUtils.pruneColumns is not called,
-            // ScanBuilder.requiredSchema is not pruned, but ScanBuilder.requiredSchema is
-            // not used anyways. The schema for aggregate columns will be built in Scan.
-            val scan = scanBuilder.build()
-
-            logInfo(
-              s"""
-                 |Pushing operators to ${relation.name}
-                 |Pushed Aggregate Functions: ${aggregation.aggregateExpressions.mkString(", ")}
-                 |Output: ${output.mkString(", ")}
-              """.stripMargin)
-            val wrappedScan = scan match {
-              case v1: V1Scan =>
-                V1ScanWrapper(v1, Seq.empty[sources.Filter], Seq.empty[sources.Filter], aggregation)
-              case _ => scan
-            }
-
-            val scanRelation = DataSourceV2ScanRelation(relation, wrappedScan, output)
-            val plan = Aggregate(groupingExpressions, resultExpressions, scanRelation)
-
-            // Change the optimized logical plan to reflect the pushed down aggregate
-            // e.g. TABLE t (c1 INT, c2 INT, c3 INT)
-            // SELECT min(c1), max(c1) FROM t;
-            // The original logical plan is
-            // Aggregate [min(c1#9) AS min(c1)#17, max(c1#9) AS max(c1)#18]
-            // +- RelationV2[c1#9] parquet ...
-            //
-            // After change the V2ScanRelation output to [min(_1)#21, max(_1)#22]
-            // we have the following
-            // !Aggregate [min(_1#9) AS min(_1)#17, max(_1#9) AS max(_1)#18]
-            // +- RelationV2[min(_1)#21, max(_1)#22] parquet ...
-            //
-            // We want to change it to
-            // == Optimized Logical Plan ==
-            // Aggregate [min(min(c1)#21) AS min(c1)#17, max(max(c1)#22) AS max(c1)#18]
-            // +- RelationV2[min(c1)#21, max(c1)#22] parquet file ...
-            var i = 0
-            plan.transformExpressions {
-              case agg: AggregateExpression =>
-                i += 1
-                val aggFunction: aggregate.AggregateFunction = agg.aggregateFunction match {
-                  case _: aggregate.Max => aggregate.Max(output(i - 1))
-                  case _: aggregate.Min => aggregate.Min(output(i - 1))
-                  case _: aggregate.Sum => aggregate.Sum(output(i - 1))
-                  case _: aggregate.Average => aggregate.Average(output(i - 1))
-                  case _: aggregate.Count => aggregate.Sum(output(i - 1))
-                  case _ => agg.aggregateFunction
-                }
-                agg.copy(aggregateFunction = aggFunction, filter = None)
-            }
-          }
-
-        case _ => aggNode // return original plan node
-      }
-
     case ScanOperation(project, filters, relation: DataSourceV2Relation) =>
       val scanBuilder = relation.table.asReadable.newScanBuilder(relation.options)
